pyproject.toml
[project]
name = "pycard"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "jinja2>=3.1.6",
    "playwright>=1.54.0",
    "ruamel-yaml>=0.18.14",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.1",
    "ruff>=0.12.8",
]

[tool.basedpyright]
reportExplicitAny = false
